+++ budget/web.py
@@ -294,7 +294,7 @@ def delete_project():
     g.project.remove_project()
     flash(_('Project successfully deleted'))
 
-    return redirect(url_for(".home"))
+    return redirect(request.headers.get('Referer') or url_for('.home'))
 
 
 @main.route("/exit")
@@ -507,5 +507,8 @@ def settle_bill():
 
 
 @main.route("/dashboard")
+@requires_admin()
 def dashboard():
-    return render_template("dashboard.html", projects=Project.query.all())
+    is_dashboard_activated = current_app.config['ACTIVATE_DASHBOARD']
+    return render_template("dashboard.html", projects=Project.query.all(),
+                           is_dashboard_activated=is_dashboard_activated)